backend/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = "This is the backend of my personal website project"
authors = ["kotova-olga <kotova.olga@bcg.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
flask = "^3.1.0"
django = "^5.1.7"
flask-cors = "^5.0.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
